requirements.txt
asgiref==3.8.1
Django==5.2
django-cors-headers==4.7.0
django-redis==5.4.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
packaging==25.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-decouple==3.8
redis==5.2.1
shortuuid==1.0.13
sqlparse==0.5.3
whitenoise==6.9.0
gunicorn==23.0.0
